Разбиение файлов по логическим блокам.
