+++ .gitignore
@@ -1,0 +1,12 @@
+*.[oa]
+*vgcore*
+.vscode/**
+*.txt
+*.out
+*.exe
+*tests*
+
+CMakeLists.txt
+cmake-build-debug/*
+.idea/*
+main.cpp
