+++ Artifact.cpp
@@ -1,4 +1,4 @@
-#include "Artifact.h"
+#include "Things.h"
 
 Artifact::Artifact(std::string name, std::string id, short level, short power) : Item(name, id, level) {
     this->power = power;
