--- Artifact.h
@@ -1,16 +1,0 @@
-#ifndef ARTIFACT_H
-#define ARTIFACT_H
-
-
-#include "Item.h"
-
-class Artifact : public Item {
-public:
-    Artifact(std::string, std::string, short, short);
-    short GetPower() const ;
-private:
-    short power;
-};
-
-
-#endif //ARTIFACT_H
