--- Hero.h
@@ -1,17 +1,0 @@
-#ifndef HERO_H
-#define HERO_H
-
-
-#include "Creature.h"
-#include "Storage.h"
-
-class Hero : public Creature {
-public:
-    Hero(std::string, std::string, short, short, coord, Storage*);
-    Storage* GetInventory() const;
-protected:
-    Storage *inventory;
-};
-
-
-#endif //HERO_H
