+++ Location.h
@@ -1,0 +1,33 @@
+#ifndef LOCALITY_H
+#define LOCALITY_H
+
+#include <vector>
+
+struct coord {
+    int x = 0;
+    int y = 0;
+    coord() = default;
+    coord(int, int);
+};
+
+class Located {
+public:
+    Located(coord);
+    void SetCoord(coord);
+    coord GetCoord() const;
+protected:
+    coord coordinates;
+};
+
+class Movable : public Located {
+public:
+    Movable(coord, short);
+    bool Move(coord);
+    coord Step();
+    short GetSpeed() const;
+protected:
+    short speed;
+    std::vector<coord> path;
+};
+
+#endif //LOCALITY_H
